Creating IndexedSequenceDataBank, for search using index. Changed the LightweightSymbolList for faster symbol list building. Changed the sequenceId and jobId for long. Created a cache for encode and decode sub sequences. A code clean up. Changed java library for version 1.5, because tptp do not suports version 1.6

+++ src/bio/pih/scheduler/AbstractWorker.java
@@ -225,7 +225,7 @@ public abstract class AbstractWorker {
 	protected class ResultSender implements Runnable {
 		private String db;
 		private String query;
-		private int code;
+		private long code;
 		private AlignmentResult[] alignmentResult;
 
 		/**
@@ -235,7 +235,7 @@ public abstract class AbstractWorker {
 		 * @param alignemtnResult
 		 *            a
 		 */
-		public ResultSender(String db, String query, int code, AlignmentResult[] alignemtnResult) {
+		public ResultSender(String db, String query, long code, AlignmentResult[] alignemtnResult) {
 			this.db = db;
 			this.query = query;
 			this.code = code;
